// src/lib/api/generated/models/v1SecurityType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1SecurityType {
    SECURITY_TYPE_UNSPECIFIED = 'SECURITY_TYPE_UNSPECIFIED',
    SECURITY_TYPE_STOCK = 'SECURITY_TYPE_STOCK',
    SECURITY_TYPE_FUTURE = 'SECURITY_TYPE_FUTURE',
    SECURITY_TYPE_FOREX = 'SECURITY_TYPE_FOREX',
    SECURITY_TYPE_OPTION = 'SECURITY_TYPE_OPTION',
}